fix/models: sha is the realm of job, not document
Each job works on a pdf (or later, a url or doi or other), and ends up
in a document, which can be an existing one (once we've implemented doi
detection to join versions of a document). Hashes are for jobs, and
later doi's (or something else if missing) will be for documents.